# docs/swagger.yaml
basePath: /api
definitions:
  handler.addPolicyRequest:
    properties:
      act:
        type: string
      dom:
        type: string
      obj:
        type: string
      sub:
        type: string
    type: object
...
  resp.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  resp.PageResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  resp.SuccessResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  service.Policy:
    properties:
      act:
        type: string
      dom:
        type: string
      id:
        type: integer
      obj:
        type: string
      ptype:
        type: string
      sub:
        type: string
    type: object
...
  service.RoleInfo:
    properties:
      name:
        type: string
      policies:
        items:
          $ref: '#/definitions/service.Policy'
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: SQL 审核、查询和工单管理平台 API
  title: SQLFlow API
  version: "1.0"
paths:
...
  /policies:
    get:
      description: 管理员获取权限策略列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 50
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 策略类型
        in: query
        name: ptype
        type: string
      - description: 主体
        in: query
        name: sub
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/resp.PageResponse'
        "500":
          description: 获取策略列表失败
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取策略列表
      tags:
      - 权限
    post:
      consumes:
      - application/json
      description: 管理员添加权限策略
      parameters:
      - description: 策略请求
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.addPolicyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: 策略添加成功
          schema:
            $ref: '#/definitions/resp.SuccessResponse'
        "400":
          description: 请求格式错误或策略已存在
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "500":
          description: 添加策略失败
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 添加策略
      tags:
      - 权限
  /policies/{id}:
    delete:
      description: 管理员删除指定权限策略
      parameters:
      - description: 策略ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 策略已删除
          schema:
            $ref: '#/definitions/resp.SuccessResponse'
        "400":
          description: 无效的策略ID
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "404":
          description: 策略不存在
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "500":
          description: 删除策略失败
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 删除策略
      tags:
      - 权限
  /policies/sync:
    post:
      description: 管理员重新加载权限策略
      produces:
      - application/json
      responses:
        "200":
          description: 策略同步成功
          schema:
            $ref: '#/definitions/resp.SuccessResponse'
        "500":
          description: 同步策略失败
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 同步策略
      tags:
      - 权限
...
  /roles:
    get:
      description: 管理员获取所有角色列表
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/resp.SuccessResponse'
      security:
      - BearerAuth: []
      summary: 获取角色列表
      tags:
      - 权限
  /roles/{role}:
    get:
      description: 管理员获取指定角色的策略信息
      parameters:
      - description: 角色名称
        in: path
        name: role
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            allOf:
            - $ref: '#/definitions/resp.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/service.RoleInfo'
              type: object
        "400":
          description: 角色名称不能为空
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
        "500":
          description: 获取角色策略失败
          schema:
            $ref: '#/definitions/resp.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取角色详情
      tags:
      - 权限
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"